Revert "ner updates"
This reverts commit f09cabb89a015cab92e62b952c7b5a124e0121ce.

+++ ADDAC_Quantizer.h
@@ -22,7 +22,7 @@
 class ADDAC_Quantizer{
 public:
 	ADDAC_Quantizer();
-	long quantize(float _val);
+	long quantize(unsigned int _val);
 	long Note(int _val, double _partial);
 	long Note(int _val);
